requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
apscheduler==3.6.3
    # via python-telegram-bot
cachetools==4.2.2
    # via python-telegram-bot
certifi==2022.6.15
    # via
    #   python-telegram-bot
    #   requests
charset-normalizer==2.0.12
    # via requests
cycler==0.11.0
    # via matplotlib
fonttools==4.34.4
    # via matplotlib
idna==3.3
    # via requests
kiwisolver==1.4.3
    # via matplotlib
matplotlib==3.5.2
    # via cryptoprophet (setup.py)
numpy==1.23.1
    # via
    #   cryptoprophet (setup.py)
    #   matplotlib
    #   pandas
    #   patsy
    #   scipy
    #   statsmodels
orjson==3.7.7
    # via cryptoprophet (setup.py)
packaging==21.3
    # via
    #   matplotlib
    #   statsmodels
pandas==1.4.3
    # via
    #   cryptoprophet (setup.py)
    #   pandas-ta
    #   statsmodels
pandas-ta==0.3.14b0
    # via cryptoprophet (setup.py)
patsy==0.5.2
    # via statsmodels
pillow==9.1.1
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.20.0
    # via cryptoprophet (setup.py)
python-telegram-bot==13.13
    # via cryptoprophet (setup.py)
pytz==2022.1
    # via
    #   apscheduler
    #   pandas
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
readchar==3.0.5
    # via cryptoprophet (setup.py)
requests==2.28.1
    # via cryptoprophet (setup.py)
scipy==1.8.1
    # via statsmodels
six==1.16.0
    # via
    #   apscheduler
    #   patsy
    #   python-dateutil
statsmodels==0.13.2
    # via cryptoprophet (setup.py)
tornado==6.1
    # via python-telegram-bot
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
urllib3==1.26.9
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
